Done: Show/hide login/out icon + prevent/allow func login/out - IMPACT

+++ src/app/component/login/login.component.ts
@@ -4,6 +4,8 @@ import { IAccount } from '../../interface/login.interface';
 import { Router } from '@angular/router';
 import { MessageService } from 'primeng/api';
 import { HttpErrorResponse } from '@angular/common/http';
+import { CommonComponent } from '../../utils/common.component';
+import { AppComponent } from '../../app.component';
 
 @Component({
   selector: 'app-login',
@@ -15,7 +17,9 @@ export class LoginComponent {
   constructor(
     private loginService: LoginService,
     private router: Router,
-    private messageService: MessageService
+    private messageService: MessageService,
+    private commmonFunc: CommonComponent,
+    private appComponent: AppComponent
   ) { }
 
   account: IAccount = {
@@ -25,22 +29,28 @@ export class LoginComponent {
   }
 
   handleLogin() {
-    this.loginService.login(this.account).subscribe({
-      next: (data) => {
-        localStorage.setItem("token", JSON.stringify(data)); //Lưu token trả về vào local storage
-        this.router.navigate(['/quanlynguoidung']);
-        this.messageService.add({ severity: 'success', summary: 'Đăng nhập', detail: 'Đăng nhập thành công!' });
-        this.clearFormRegister();
-      },
-      error: (error: HttpErrorResponse) => {
-        if (error.error) {
-          const fieldErrors = error.error.result;
-          if (fieldErrors.lockedAccount) {
-            this.messageService.add({ severity: 'error', summary: 'Lỗi', detail: `${fieldErrors.lockedAccount}` });
+    const isExistTokenOrLoggedIn = this.commmonFunc.checkUserRoleOrLoggedIn();
+    if (isExistTokenOrLoggedIn) {
+      this.messageService.add({ severity: 'error', summary: 'Lỗi', detail: "Bạn đã đăng nhập rồi!" });
+    } else {
+      this.loginService.login(this.account).subscribe({
+        next: (data) => {
+          localStorage.setItem("token", JSON.stringify(data)); //Lưu token trả về vào local storage
+          this.router.navigate(['/quanlynguoidung']);
+          this.appComponent.checkLoginStatus();
+          this.messageService.add({ severity: 'success', summary: 'Đăng nhập', detail: 'Đăng nhập thành công!' });
+          this.clearFormRegister();
+        },
+        error: (error: HttpErrorResponse) => {
+          if (error.error) {
+            const fieldErrors = error.error.result;
+            if (fieldErrors.lockedAccount) {
+              this.messageService.add({ severity: 'error', summary: 'Lỗi', detail: `${fieldErrors.lockedAccount}` });
+            }
           }
         }
-      }
-    })
+      })
+    }
   }
 
   clearFormRegister() {
